refactor tests: extract generateAndTest to utils
